Module liftof_tui::tab_commands

Commands tab

Display available commands as a list, if issued show the corresponding response recieved over the command channel and show the incoming data stream as well

The layout is somewhat like this

| Menu | .. | .. | |

| Command List | | Response: | | .. | | ––––––––––– | | | Stream | | | | -> … |

| Logs |

Structs

CommandTab