Crate liftof_tui

Re-exports

pub use crate::tabs::*;

pub use crate::layout::*;

Modules

colors

layout

Main laoyt of the app

menu

Menus for the main application and individual tabs

tabs

widgets

Functions

packet_distributor

Receive packets from an incoming stream and distrubute them to their receivers while taking notes of everything

render_logs

Use the TuiLoggerWidget to display the most recent log messages

telly_packet_counter

A map which keeps track of the types of telemetry packets received