liftof_tui
Module colors

Structs

ColorSet

Implementation of a color palette

ColorTheme

A color theme, created from a color palette whcih allows to provide style variants for ui elements

Constants

COLORSETBW

Black and white color palette for highest contrast

COLORSETDUNE

A color palette inspired by the recent Dune movie

COLORSETGAPS

A color palette created from the colors of the GAPS logo

COLORSETLD

A color palette inspired by Star Trek Lower Decks

COLORSETMATRIX

A color palette inspired by the original Matrix trilogy

COLORSETNIUHI

A color palette designed for the NIUHI brnach

COLORSETOMILU

A color palette designed for the OMILU branch

COLORSETPRINCESSPEACH

A pink color palette

COLORSETSTARFIELD

A color palette inspired by Bethesda’s recent ARPG