liftof_tui

Function packet_distributor

pub fn packet_distributor(
    tp_from_sock: Receiver<TofPacket>,
    tp_sender_mt: Sender<TofPacket>,
    tp_sender_rb: Sender<TofPacket>,
    tp_sender_ev: Sender<TofPacket>,
    tp_sender_cp: Sender<TofPacket>,
    tp_sender_tr: Sender<TofPacket>,
    rbwf_sender: Sender<TofPacket>,
    ts_send: Sender<TofPacket>,
    th_send: Sender<TofHit>,
    tp_sender_hb: Sender<TofPacket>,
    str_list: Arc<Mutex<VecDeque<String>>>,
    pck_map: Arc<Mutex<HashMap<&str, usize>>>,
    writer: Option<TofPacketWriter>,
)

Receive packets from an incoming stream and distrubute them to their receivers while taking notes of everything

This is a Pablo Pubsub kind of persona (see a fantastic talk at RustConf 2023)