Module liftof_tui::tabs

Re-exports

pub use crate::tab_mt::MTTab;

pub use crate::tab_settings::SettingsTab;

pub use crate::tab_home::HomeTab;

pub use crate::tab_events::EventTab;

pub use crate::tab_paddles::PaddleTab;

pub use crate::tab_heartbeats::HeartBeatTab;

pub use crate::tab_heartbeats::HeartBeatView;

pub use crate::tab_rbs::RBTab;

pub use crate::tab_rbs::RBTabView;

pub use crate::tab_rbs::RBLTBListFocus;

pub use crate::tab_telemetry::TelemetryTab;

pub use crate::tab_rbwaveform::RBWaveformTab;

pub use crate::tab_tofsummary::TofSummaryTab;

pub use crate::tab_tofhit::TofHitTab;

pub use crate::tab_tofhit::TofHitView;

pub use crate::tab_cpu::CPUTab;

pub use crate::tab_commands::CommandTab;

Modules

tab_commands

Commands tab

tab_cpu

tab_events

tab_heartbeats

Heartbeats are a special kind of monitoring which monitor the individual main threads of liftof-cc and are sent in regular intervals

tab_home

The front page of the app, including a view of received packets

tab_mt

Master Trigger tab

tab_paddles

tab_rbs

ReadoutBoard Status tab

tab_rbwaveform

tab_settings

tab_telemetry

tab_tofhit

tab_tofsummary