liftof_tui

Module widgets

Constants

LG_LINE

LG_LINE_HORIZONTAL

Functions

clean_data

Adapt the bins of the histogram for the bar chart which will get rendered. Always show a minimum number of bins, but if the max y-bin is “too far to the left” then shorten the range for a better visualization

create_labels

Create the labels for a certain histogram for rendering

gauge

A simple line gauge, that is bacically a progress bar

histogram

prep_data

Prepare data for histogram widget

timeseries